Allow resuming from a previous chainsate - CHAINSTATE_DIR=... make up with an absoluate path will resume a network using   that dir. - will not work with the genesis chainstate dir, and if a relative path   is used, a genesis sync will start (using an epoch dir) from a nested dir like ./docker/docker/chainstate/<epoch>

+++ Makefile
@@ -30,19 +30,22 @@ STRESS_TIMEOUT ?= 120
 
 # Create the chainstate dir and extract an archive to it when the "up" target is used
 $(CHAINSTATE_DIR): /usr/bin/tar /usr/bin/zstd
-	@mkdir -p $(CHAINSTATE_DIR)
-	@if [ "$(TARGET)" = "up" ]; then
-		if [ -f "$(CHAINSTATE_ARCHIVE)" ]; then
-			sudo tar --same-owner -xf $(CHAINSTATE_ARCHIVE) -C $(CHAINSTATE_DIR) || exit 1
-		else
-			@echo "Chainstate archive ($(CHAINSTATE_ARCHIVE)) not found. Exiting"
-			rm -rf $(CHAINSTATE_DIR)
-			exit 1
+	@if  [ ! -d "$(CHAINSTATE_DIR)" ]; then \
+		@mkdir -p $(CHAINSTATE_DIR)
+		@if [ "$(TARGET)" = "up" ]; then
+			if [ -f "$(CHAINSTATE_ARCHIVE)" ]; then
+				sudo tar --same-owner -xf $(CHAINSTATE_ARCHIVE) -C $(CHAINSTATE_DIR) || exit 1
+			else
+				@echo "Chainstate archive ($(CHAINSTATE_ARCHIVE)) not found. Exiting"
+				rm -rf $(CHAINSTATE_DIR)
+				exit 1
+			fi
 		fi
 	fi
 
 # Boot the network from the local chainstate archive
-up: check-not-running | build $(CHAINSTATE_DIR)
+up: check-not-running | $(CHAINSTATE_DIR)
+# up: check-not-running | build $(CHAINSTATE_DIR)
 	@echo "Starting $(PROJECT) network from chainstate archive"
 	@echo "  Chainstate Dir: $(CHAINSTATE_DIR)"
 	@echo "  Chainstate Archive: $(CHAINSTATE_ARCHIVE)"
